tests on the front and backend are working
