+++ grails-app/services/org/bbop/apollo/FeatureService.groovy
@@ -1877,44 +1877,9 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
 
 
 
-    // TODO: add unit tests and move to a non-transcration method
-    @NotTransactional
-    def getSOUrlForCvTermLabels(List<String> cvTermsArray){
-        return cvTermsArray.collect{  getSOUrlForCvTerm(it)}.findAll{ it!=null}
-    }
-
-    @NotTransactional
-    String getSOUrlForCvTerm(String cvTerm){
-        String ontologyId = null
-        try {
-            ontologyId = Class.forName("org.bbop.apollo.feature.${cvTerm}")?.ontologyId
-        } catch (e) {
-        }
-        try {
-            ontologyId = ontologyId ?: Class.forName("org.bbop.apollo.variant.${cvTerm}")?.ontologyId
-        } catch (e) {
-        }
-        return ontologyId
-    }
-
-    String findMostSpecificLabel(Collection<String> labels) {
-        def filteredLabels = labels.findAll { it != "Feature" && !it.contains("TranscriptRegion") && it != "SpliceSite" }
-        if(filteredLabels.contains("MRNA")) return "MRNA"
-        log.debug "filtered labels ${filteredLabels}"
-//        if (filteredLabels.indexOf("Transcript") >= 0) {
-//            if (filteredLabels.size() > 1) {
-//                return filteredLabels.findAll { it != "Transcript" }.first()
-//            } else {
-//                return filteredLabels.last()
-//            }
-//        }
-        return filteredLabels.last()
-
-    }
-
     String getCvTermFromNeo4jFeature(def feature) {
         log.debug "cv term ${feature}"
-        String specificType = findMostSpecificLabel(feature.labels())
+        String specificType = FeatureTypeMapper.findMostSpecificLabel(feature.labels())
         log.debug "specific type ${specificType}"
         return Class.forName("org.bbop.apollo.feature.${specificType}")?.cvTerm
     }
@@ -1935,66 +1900,6 @@ public void setTranslationEnd(Transcript transcript, int translationEnd) {
     }
 
 
-    @NotTransactional
-    def castNeo4jFeature(InternalNode internalNode) {
-        String label = findMostSpecificLabel(internalNode.labels())
-        switch (label.toUpperCase()) {
-            case MRNA.cvTerm.toUpperCase(): return internalNode as MRNA
-            case MiRNA.cvTerm.toUpperCase(): return internalNode as MiRNA
-            case NcRNA.cvTerm.toUpperCase(): return internalNode as NcRNA
-            case GuideRNA.cvTerm.toUpperCase(): return internalNode as GuideRNA
-            case RNasePRNA.cvTerm.toUpperCase(): return internalNode as RNasePRNA
-            case TelomeraseRNA.cvTerm.toUpperCase(): return internalNode as TelomeraseRNA
-            case SrpRNA.cvTerm.toUpperCase(): return internalNode as SrpRNA
-            case LncRNA.cvTerm.toUpperCase(): return internalNode as LncRNA
-            case RNaseMRPRNA.cvTerm.toUpperCase(): return internalNode as RNaseMRPRNA
-            case ScRNA.cvTerm.toUpperCase(): return internalNode as ScRNA
-            case PiRNA.cvTerm.toUpperCase(): return internalNode as PiRNA
-            case TmRNA.cvTerm.toUpperCase(): return internalNode as TmRNA
-            case EnzymaticRNA.cvTerm.toUpperCase(): return internalNode as EnzymaticRNA
-            case SnoRNA.cvTerm.toUpperCase(): return internalNode as SnoRNA
-            case SnRNA.cvTerm.toUpperCase(): return internalNode as SnRNA
-            case RRNA.cvTerm.toUpperCase(): return internalNode as RRNA
-            case TRNA.cvTerm.toUpperCase(): return internalNode as TRNA
-            case Transcript.cvTerm.toUpperCase(): return internalNode as Transcript
-            case Gene.cvTerm.toUpperCase(): return internalNode as Gene
-            case Exon.cvTerm.toUpperCase(): return internalNode as Exon
-            case CDS.cvTerm.toUpperCase(): return internalNode as CDS
-            case Intron.cvTerm.toUpperCase(): return internalNode as Intron
-            case Pseudogene.cvTerm.toUpperCase(): return internalNode as Pseudogene
-            case PseudogenicRegion.cvTerm.toUpperCase(): return internalNode as PseudogenicRegion
-            case ProcessedPseudogene.cvTerm.toUpperCase(): return internalNode as ProcessedPseudogene
-            case TransposableElement.alternateCvTerm.toUpperCase():
-            case TransposableElement.cvTerm.toUpperCase(): return internalNode as TransposableElement
-            case Terminator.alternateCvTerm.toUpperCase():
-            case Terminator.cvTerm.toUpperCase(): return internalNode as Terminator
-            case ShineDalgarnoSequence.alternateCvTerm.toUpperCase():
-            case ShineDalgarnoSequence.cvTerm.toUpperCase(): return internalNode as ShineDalgarnoSequence
-            case RepeatRegion.alternateCvTerm.toUpperCase():
-            case RepeatRegion.cvTerm.toUpperCase(): return internalNode as RepeatRegion
-            case InsertionArtifact.cvTerm.toUpperCase(): return internalNode as InsertionArtifact
-            case DeletionArtifact.cvTerm.toUpperCase(): return internalNode as DeletionArtifact
-            case SubstitutionArtifact.cvTerm.toUpperCase(): return internalNode as SubstitutionArtifact
-            case Insertion.cvTerm.toUpperCase(): return internalNode as Insertion
-            case Deletion.cvTerm.toUpperCase(): return internalNode as Deletion
-            case Substitution.cvTerm.toUpperCase(): return internalNode as Substitution
-            case StopCodonReadThrough.cvTerm.toUpperCase(): return internalNode as StopCodonReadThrough
-            case NonCanonicalFivePrimeSpliceSite.cvTerm.toUpperCase(): return internalNode as NonCanonicalFivePrimeSpliceSite
-            case NonCanonicalThreePrimeSpliceSite.cvTerm.toUpperCase(): return internalNode as NonCanonicalThreePrimeSpliceSite
-            case NonCanonicalFivePrimeSpliceSite.alternateCvTerm.toUpperCase(): return internalNode as NonCanonicalFivePrimeSpliceSite
-            case NonCanonicalThreePrimeSpliceSite.alternateCvTerm.toUpperCase(): return internalNode as NonCanonicalThreePrimeSpliceSite
-            case SNV.cvTerm.toUpperCase(): return internalNode as SNV
-            case SNP.cvTerm.toUpperCase(): return internalNode as SNP
-            case MNV.cvTerm.toUpperCase(): return internalNode as MNV
-            case MNP.cvTerm.toUpperCase(): return internalNode as MNP
-            case Indel.cvTerm.toUpperCase(): return internalNode as Indel
-            default:
-                log.error("None for type: " + internalNode)
-//        log.error("No feature type exists for ${ontologyId}")
-                return internalNode as Feature
-        }
-
-    }
 
     // TODO: (perform on client side, slightly ugly)
     Feature generateFeatureForType(String ontologyId) {
